refactor: move squad worker checkout helpers into their own module
